@@ -1,11 +1,17 @@
-module gitlab.com/sparetimecoders/goamqp
+module github.com/sparetimecoders/goamqp
 
-go 1.14
+go 1.18
 
 require (
 	github.com/caarlos0/env v3.5.0+incompatible
-	github.com/google/uuid v1.1.1
+	github.com/google/uuid v1.3.0
 	github.com/pkg/errors v0.9.1
-	github.com/streadway/amqp v0.0.0-20190827072141-edfb9018d271
-	github.com/stretchr/testify v1.3.0
+	github.com/rabbitmq/amqp091-go v1.3.4
+	github.com/stretchr/testify v1.8.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
