@@ -1,3 +1,25 @@
+// MIT License
+//
+// Copyright (c) 2019 sparetimecoders
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
 package goamqp
 
 import (
@@ -12,7 +34,7 @@ type MessageLogger func(jsonContent []byte, eventType reflect.Type, routingKey s
 
 // NoOpMessageLogger is a MessageLogger that will do nothing
 // This is the default implementation if the setup func UseMessageLogger is not used
-func NoOpMessageLogger() MessageLogger {
+func noOpMessageLogger() MessageLogger {
 	return func(jsonContent []byte, eventType reflect.Type, routingKey string, outgoing bool) {
 	}
 }
@@ -26,7 +48,7 @@ func StdOutMessageLogger() MessageLogger {
 		if err != nil {
 			prettyJSONString = string(jsonContent)
 		} else {
-			prettyJSONString = string(prettyJSON.Bytes())
+			prettyJSONString = prettyJSON.String()
 		}
 		if outgoing {
 			fmt.Printf("Sending [%s] using routingkey: '%s' with content:\n%s\n", eventType, routingKey, prettyJSONString)
